backend/requirements.txt
# FastAPI & Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
websockets==12.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
aiosqlite==0.19.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Celery & Redis
celery==5.3.6
redis==5.0.1
flower==2.0.1

# PDF Processing - Core
PyMuPDF==1.23.8
pdf2image==1.16.3
pdfplumber==0.10.3
reportlab==4.0.8
pypdf==3.17.4
pikepdf==8.10.1

# Document Processing
Pillow==10.2.0
python-docx==1.1.0
openpyxl==3.1.2
python-pptx==0.6.23
markdown==3.5.2
beautifulsoup4==4.12.2
lxml==5.1.0
camelot-py[cv]==0.11.0
tabula-py==2.8.2
pandas==2.1.4

# OCR
pytesseract==0.3.10
pdf2image==1.16.3

# AI/ML
openai==1.10.0
anthropic==0.12.0
tiktoken==0.5.2
langchain==0.1.4
langchain-openai==0.0.5
sentence-transformers==2.2.2

# Cloud Storage
boto3==1.34.14
google-cloud-storage==2.14.0
dropbox==11.36.2

# Image Processing
opencv-python-headless==4.9.0.80
numpy==1.26.3
scikit-image==0.22.0

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.26.0
jinja2==3.1.3
python-dateutil==2.8.2
qrcode==7.4.2
PyJWT==2.8.0
cryptography==41.0.7

# Data Validation & Processing
email-validator==2.1.0
phonenumbers==8.13.27
pycountry==23.12.11

# Compression
zstandard==0.22.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

